Remove some functions and fix a helper; #1053
